Build dashboard with binary resource bundle
