+++ src/oggex/oggex.gresource.xml.in
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<gresources>
+    <gresource prefix="/com/github/jmd/dashboard">
+        <file preprocess="xml-stripblanks">dashboard.glade</file>
+  </gresource>
+</gresources>
+
