+++ src/oggex/resources
@@ -1,0 +1,1 @@
+/home/jmd/workspace/apps/oggex/resources/
